flyttet ManglerTilgangLenkePanel ut i egen fil, la til lenker
